Move Data codec to services module
