+++ services/src/test/java/io/scalecube/services/StreamDataCodecTest.java
@@ -1,6 +1,9 @@
-package io.scalecube.streams;
+package io.scalecube.services;
 
-import io.scalecube.streams.codec.StreamMessageDataCodecImpl;
+import io.scalecube.streams.ClientStream;
+import io.scalecube.streams.Event;
+import io.scalecube.streams.ListeningServerStream;
+import io.scalecube.streams.StreamMessage;
 import io.scalecube.streams.codec.StreamMessageDataCodec;
 import io.scalecube.transport.Address;
